Refactor main-product into mp--media and mp--info child architecture
Split the flat main-product section into two Weaverse child components:
- mp--media: wraps ProductMedia with all media settings, hosts badges as overlay
- mp--info: sticky container for all product info children (breadcrumb, title, etc.)

This moves badges onto the product media as an absolute-positioned overlay,
eliminating layout shift when badges appear/disappear on variant change.
