+++ app/sections/main-product/product-badges.tsx
@@ -31,7 +31,7 @@ export default function ProductBadgesComponent(
     <div
       ref={ref}
       {...rest}
-      className="flex items-center gap-2 text-sm empty:hidden"
+      className="absolute top-2 left-2 z-1 flex items-center gap-2 text-sm empty:hidden md:top-4 md:left-4"
     >
       <ProductBadges
         as={Fragment}
